Refactor game logic into modules
