+++ pieces.js
@@ -1,0 +1,147 @@
+export const SHAPES = {
+  I: [
+    [
+      [0, 0, 0, 0],
+      [1, 1, 1, 1],
+      [0, 0, 0, 0],
+      [0, 0, 0, 0],
+    ],
+    [
+      [0, 0, 1, 0],
+      [0, 0, 1, 0],
+      [0, 0, 1, 0],
+      [0, 0, 1, 0],
+    ],
+  ],
+  O: [
+    [
+      [0, 1, 1, 0],
+      [0, 1, 1, 0],
+      [0, 0, 0, 0],
+      [0, 0, 0, 0],
+    ],
+  ],
+  T: [
+    [
+      [0, 1, 0],
+      [1, 1, 1],
+      [0, 0, 0],
+    ],
+    [
+      [0, 1, 0],
+      [0, 1, 1],
+      [0, 1, 0],
+    ],
+    [
+      [0, 0, 0],
+      [1, 1, 1],
+      [0, 1, 0],
+    ],
+    [
+      [0, 1, 0],
+      [1, 1, 0],
+      [0, 1, 0],
+    ],
+  ],
+  S: [
+    [
+      [0, 1, 1],
+      [1, 1, 0],
+      [0, 0, 0],
+    ],
+    [
+      [0, 1, 0],
+      [0, 1, 1],
+      [0, 0, 1],
+    ],
+  ],
+  Z: [
+    [
+      [1, 1, 0],
+      [0, 1, 1],
+      [0, 0, 0],
+    ],
+    [
+      [0, 0, 1],
+      [0, 1, 1],
+      [0, 1, 0],
+    ],
+  ],
+  J: [
+    [
+      [1, 0, 0],
+      [1, 1, 1],
+      [0, 0, 0],
+    ],
+    [
+      [0, 1, 1],
+      [0, 1, 0],
+      [0, 1, 0],
+    ],
+    [
+      [0, 0, 0],
+      [1, 1, 1],
+      [0, 0, 1],
+    ],
+    [
+      [0, 1, 0],
+      [0, 1, 0],
+      [1, 1, 0],
+    ],
+  ],
+  L: [
+    [
+      [0, 0, 1],
+      [1, 1, 1],
+      [0, 0, 0],
+    ],
+    [
+      [0, 1, 0],
+      [0, 1, 0],
+      [0, 1, 1],
+    ],
+    [
+      [0, 0, 0],
+      [1, 1, 1],
+      [1, 0, 0],
+    ],
+    [
+      [1, 1, 0],
+      [0, 1, 0],
+      [0, 1, 0],
+    ],
+  ],
+};
+
+export const COLORS = {
+  I: "#38bdf8",
+  O: "#facc15",
+  T: "#a855f7",
+  S: "#22c55e",
+  Z: "#f97316",
+  J: "#3b82f6",
+  L: "#f59e0b",
+};
+
+export class Bag {
+  constructor() {
+    this.queue = [];
+    this.refill();
+  }
+
+  refill() {
+    const types = Object.keys(SHAPES);
+    for (let i = types.length - 1; i > 0; i -= 1) {
+      const j = Math.floor(Math.random() * (i + 1));
+      [types[i], types[j]] = [types[j], types[i]];
+    }
+    this.queue.push(...types);
+  }
+
+  draw() {
+    if (this.queue.length === 0) {
+      this.refill();
+    }
+    return this.queue.pop();
+  }
+}
